requirements.txt
# 核心依赖
pandas>=1.5.0
requests>=2.28.0
watchdog>=2.1.0
openpyxl>=3.0.0
Pillow>=9.0.0
urllib3>=1.26.0

# GUI界面依赖
PyQt5>=5.15.0
